cpumask: add unit tests
Adding unit tests to cpumask was made difficult by AddressSanitizer's objection
that passing a const char* pointer from C++ to C results in a use-after-scope
error.  The complaint is odd since the same sequence of operations in C in
cpumask.c itself does not trigger ASAN.  Inserting "extern C" in logical places
did not help.  Wrapping the const char* pointer in a std::shared_ptr suppressed
the use-after-scope error, but resulted in ASAN SEGV-ing, presumably because the
C function parse_range() freed the raw pointer that it was passed from C++.