cpumask: see if ASAN works with Catch2 test framework
The Catch2 test framework has some limitations compared to googletest in that it
exits when the tested program does and therefore cannot observe program exit
codes.  However, it does not trigger an use-after-scope error when it observes
the parse_range() function execution.  ASAN does still SEGV when observing the
shared_ptr version of the code, however.