classify_process_affinity_lib: determine if Linux scheduler will move a thread
Inspired by Carsten Emde's patch in

	 https://lore.kernel.org/util-linux/20220805154820.407972387@osadl.org/T/#t

read stat files from procfs and parse them to determine if the CPU affinity of a
given thread is settable or not.  Return a std::set containing the names and
PIDs of all threads along with the affinity settability for each.

Add clang-tidy, GCOV and valgrind tests to the Makefile in order to
check the code.