classify_process_affinity: fix broken unit test
A directory iterator presents file in /proc based on their lexigraphical order,
but the tid_data set in which the file contents are stored using the thread_name
as its key.  Therefore if the TID of a thread changes, the TID stored in the set
may differ from that in procfs.     The reason why the existing test failed is
due to differing TIDs.

Given that the purpose of classify_process_affinity is to mark which threads are
pinnable, the TID is uninteresting anyway.  It's because thread names are
durable and TIDs are not that thread_names are chosen as the set key in the
first place.