timerlat_pipe_load: pass responding function as a parameter
Pass the function that writes to the FIFO as a parameter to the startup code.
Subsequent attempts to craft a function that would fail to start without causing
the test to exit have not been successful.   I need to modify the code so that
it does not abort when passed the empty function:

$ gdb ./timerlat_pipe_load_lib_test
(gdb) run

terminate called after throwing an instance of 'std::bad_function_call'
  what():  bad_function_call

Thread 5 "timerlat_pipe_l" received signal SIGABRT, Aborted.
[Switching to Thread 0x7ffff2e006c0 (LWP 2599168)]
__pthread_kill_implementation (threadid=<optimized out>, signo=signo@entry=6, no_tid=no_tid@entry=0)